Added some roles and editing posts for admin

+++ app/Http/Controllers/ArticleController.php
@@ -9,39 +9,43 @@ use App\Http\Controllers\Controller;
 
 class ArticleController extends Controller
 {
-    public function index(){
-        $articles=Article::all();
-        return view('articles',['articles'=>$articles]);
+    public function index()
+    {
+        $articles = Article::all();
+        return view('articles', ['articles' => $articles]);
 
     }
-    public function show($id){
+
+    public function show($id)
+    {
         $this->authorize('show');
-        $article=Article::find($id);
-        return view('show',['article'=>$article]);
+        $article = Article::find($id);
+        return view('show', ['article' => $article]);
     }
-    public function create(){
+
+    public function create()
+    {
         $this->authorize('create');
         return view('create');
     }
-    public function store(Request $request){
+
+    public function store(Request $request)
+    {
         Article::create($request->all());
         return redirect('/articles');
     }
 
-    public function edit($id){
-        $article=Article::find($id);
-        return view('edit',['article'=>$article]);
+    public function edit($id)
+    {
+        $this->authorize('edit');
+        $article = Article::find($id);
+        return view('edit', ['article' => $article]);
     }
 
-    public function update(Request $request,$id){
-        $article=Article::find($id);
-        $article->title=$request->title;
-        $article->content=$request->content;
-        $article->save();
-        return redirect('/articles');
-    }
-    public function delete($id){
-        $article=Article::find($id);
+    public function delete($id)
+    {
+        $this->authorize('delete');
+        $article = Article::find($id);
         $article->delete();
         return redirect('/articles');
